reemplazar viejo TypeChecker por el nuevo
